+++ cppcache/src/CqEventImpl.cpp
@@ -100,8 +100,9 @@ std::string CqEventImpl::toString() {
   ACE_OS::snprintf(
       buffer, 1024,
       "CqEvent CqName=%s; base operation=%d; cq operation= %d;key=%s;value=%s",
-      m_cQuery->getName().c_str(), m_baseOp, m_queryOp,
-      m_key->toString().c_str(), m_newValue->toString().c_str());
+      m_cQuery->getName().c_str(), static_cast<int>(m_baseOp),
+      static_cast<int>(m_queryOp), m_key->toString().c_str(),
+      m_newValue->toString().c_str());
   return buffer;
 }
 std::shared_ptr<CacheableBytes> CqEventImpl::getDeltaValue() const {
