+++ cppcache/src/PdxFieldType.cpp
@@ -137,8 +137,8 @@ std::string PdxFieldType::toString() const {
   ACE_OS::snprintf(
       stringBuf, 1024,
       " PdxFieldName=%s TypeId=%d VarLenFieldIdx=%d sequenceid=%d\n",
-      this->m_fieldName.c_str(), this->m_typeId, this->m_varLenFieldIdx,
-      this->m_sequenceId);
+      this->m_fieldName.c_str(), static_cast<int>(this->m_typeId),
+      this->m_varLenFieldIdx, this->m_sequenceId);
   return std::string(stringBuf);
 }
